Crate wit_parser

Modules

abi

Structs

Case

Docs

Enum

EnumCase

Field

Flag

Flags

Function

Interface

LiveTypes

Package

A WIT package within a Resolve.

PackageName

A structure used to keep track of the name of a package, containing optional information such as a namespace and version information.

Record

Remap

Structure returned by Resolve::merge which contains mappings from old-ids to new-ids after the merge.

Resolve

Representation of a fully resolved set of WIT packages.

Result_

SizeAlign

SourceMap

A listing of source files which are used to get parsed into an UnresolvedPackage.

Stream

Tuple

TypeDef

Union

UnionCase

UnresolvedPackage

Representation of a parsed WIT package which has not resolved external dependencies yet.

Variant

World

Enums

FlagsRepr

FunctionKind

Int

Results

ResultsTypeIter

Type

TypeDefKind

TypeOwner

WorldItem

WorldKey

The key to the import/export maps of a world. Either a kebab-name or a unique interface.

Functions

validate_id

Checks if the given string is a legal identifier in wit.

Type Aliases

InterfaceId

PackageId

Params

TypeId

WorldId